Move CoalescedLocation to new file
